Add Enhancement-37 (operator-correctness audit + fixes)
A systematic operator-correctness audit covering the arithmetic, relational,
logical, bitwise/shift, ternary and concatenation families: 60+ checks in five
self-checking modules, each failing check contributing a distinct power of two
to a score emitted on a signal-flow output (v(out) == 0 means all pass; nonzero
is a bitmask pinpointing the failing check). Three real defects found and fixed;
everything else checked out exactly correct.

- hir_lower/src/expr.rs: `~x` (bitwise NOT) was lowered as ARITHMETIC NEGATION
  (`ineg`): ~12 evaluated to -12 instead of -13 -- silent wrong answers in any
  bit-manipulating model. Fixed to `inot` (whose constant fold, `!val`, was
  already correct; only the lowering picked the wrong opcode).

- mir_opt/src/const_eval.rs: constant folding of `>>` sign-extended. The MIR and
  the LLVM runtime path correctly distinguish logical Ishr/LShr (`>>`) from
  arithmetic Iashr/AShr (`>>>`), but the constant folder computed BOTH with
  Rust's signed `>>`: -16 >> 2 folded to -4 instead of the zero-filled
  1073741820, and a constant `>>` disagreed with the identical runtime
  expression. Fixed by folding through u32.

- hir_ty/src/types.rs: the ternary operator rejected string operands
  (cond ? "a" : "b" was a type error). Fixed by appending a
  (String, String) -> String signature to the SELECT list (appended last so the
  existing signature indices stay stable); the phi lowering handled strings
  unchanged.

Verified (examples/operator_examples/): the audit compiles (string ternaries
used to error), all five family scores read exactly 0, and the three
formerly-broken cases are asserted directly (~12 == -13,
-16 >> 2 == 1073741820, string ternary). Regressions: all 33 example verify
suites pass and all 75 example models recompile.

+++ OpenVAF-master-20260610/openvaf/mir_opt/src/const_eval.rs
@@ -12,7 +12,11 @@ pub fn eval_binary(func: &mut Function, op: Opcode, lhs: Const, rhs: Const) -> V
             Opcode::Irem => func.dfg.iconst(lhs % rhs),
 
             Opcode::Ishl => func.dfg.iconst(lhs << rhs),
-            Opcode::Ishr => func.dfg.iconst(lhs >> rhs),
+            // Enhancement-37: `Ishr` is the LOGICAL (zero-fill) right shift (`>>`);
+            // Rust's `>>` on a signed i32 sign-extends, which is `Iashr`'s (`>>>`)
+            // semantics -- fold through u32 so the vacated bits are zero-filled
+            // (matches the LLVMBuildLShr the runtime path emits).
+            Opcode::Ishr => func.dfg.iconst(((lhs as u32) >> rhs) as i32),
             // `>>` on a signed value is already an arithmetic (sign-extending) shift.
             Opcode::Iashr => func.dfg.iconst(lhs >> rhs),
             Opcode::Ixor => func.dfg.iconst(lhs ^ rhs),
